feat: implement dateUtils
