+++ packages/model/src/index.ts
@@ -1,3 +1,4 @@
 export * from './messages'
 export * from './events'
 export * from './types'
+export * from './utils'
