+++ packages/model/src/messages/mrtd/EMrtdDataSubmitMessage.ts
@@ -1,6 +1,7 @@
 import { EMrtdData } from '@2060.io/credo-ts-didcomm-mrtd'
 import * as Mrz from 'mrz'
 
+import { convertShortDate } from '../../utils'
 import { BaseMessage, BaseMessageOptions } from '../BaseMessage'
 import { MessageType } from '../MessageType'
 
@@ -99,7 +100,7 @@ export class EMrtdDataSubmitMessage extends BaseMessage {
         documentType: parsedMrz.format,
         documentNumber: parsedMrz.documentNumber ?? undefined,
         issuingState: parsedMrz.fields.issuingState ?? undefined,
-        dateOfExpiry: this.convertMRTDDate(parsedMrz.fields.expirationDate, true),
+        dateOfExpiry: convertShortDate(parsedMrz.fields.expirationDate, true),
         sex: parsedMrz.fields.sex ?? undefined,
         nationality: parsedMrz.fields.nationality ?? undefined,
         lastName: parsedMrz.fields.lastName ?? undefined,
@@ -109,7 +110,7 @@ export class EMrtdDataSubmitMessage extends BaseMessage {
         nameOfHolder:
           parsed.fields.additionalPersonalData?.nameOfHolder ??
           `${parsedMrz.fields.lastName} ${parsedMrz.fields.firstName}`,
-        dateOfBirth: this.convertMRTDDate(dateOfBirth, false),
+        dateOfBirth: convertShortDate(dateOfBirth, false),
         otherNames: parsed.fields.additionalPersonalData?.otherNames,
         personalNumber: parsed.fields.additionalPersonalData?.personalNumber,
         placeOfBirth: parsed.fields.additionalPersonalData?.placeOfBirth,
@@ -122,54 +123,5 @@ export class EMrtdDataSubmitMessage extends BaseMessage {
       },
     }
     return newEmrtdData
-  }
-
-  /**
-   * Converts a Machine Readable Travel Document (MRTD) date in the format `YYMMDD` to a complete
-   * `YYYYMMDD` date format, taking into account the current century.
-   *
-   * **Note:** This method is limited to interpreting dates based on the current year.
-   * It may not handle dates correctly for years beyond the range determined by the
-   * current century (e.g., for dates after 2050 when the current year is in the 21st century).
-   *
-   * @param {string} date - The MRTD date string in the format `YYMMDD`.
-   * @param {boolean} isExpirationDate - A boolean flag indicating whether the date is an expiration date.
-   * @returns {string} - The converted date in the format `YYYYMMDD`, or the original input
-   *                     if the input is not a valid `YYMMDD` date.
-   *
-   * @example
-   * // Current year: 2024
-   * convertMRTDDate("240101"); // Returns "20240101"
-   * convertMRTDDate("991231"); // Returns "19991231"
-   * convertMRTDDate("abcd12"); // Returns "abcd12" (invalid input)
-   */
-  public convertMRTDDate(date: string | null | undefined, isExpirationDate: boolean) {
-    if (!date || !/^\d{6}$/.test(date)) return date ?? undefined
-
-    const currentYear = new Date().getFullYear()
-    const currentCentury = Math.floor(currentYear / 100)
-    const year = parseInt(date.slice(0, 2), 10)
-    const month = date.slice(2, 4)
-    const day = date.slice(4, 6)
-
-    let fullYear: number
-
-    if (isExpirationDate) {
-      if (year <= currentYear % 100) {
-        fullYear = currentCentury * 100 + year
-        if (fullYear < currentYear) {
-          fullYear += 100
-        }
-      } else {
-        fullYear = currentCentury * 100 + year
-      }
-    } else {
-      if (year <= currentYear % 100) {
-        fullYear = currentCentury * 100 + year
-      } else {
-        fullYear = (currentCentury - 1) * 100 + year
-      }
-    }
-    return `${fullYear}${month}${day}`
   }
 }
